# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.City:
    properties:
      alias:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      province:
        $ref: '#/definitions/models.Province'
      province_id:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Country:
    properties:
      alias:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      provinces:
        items:
          $ref: '#/definitions/models.Province'
        type: array
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.Hotel:
    properties:
      address:
        type: string
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      email_address:
        type: string
      extra_data:
        type: string
      fax_number:
        type: string
      hotel_grade:
        $ref: '#/definitions/models.HotelGrade'
      hotel_grade_id:
        type: integer
      hotel_type:
        $ref: '#/definitions/models.HotelType'
      hotel_type_id:
        type: integer
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        type: integer
      phone_number1:
        type: string
      phone_number2:
        type: string
      postal_code:
        type: string
      province:
        $ref: '#/definitions/models.Province'
      province_id:
        type: integer
      tenant_id:
        type: integer
      thumbnails:
        items:
          $ref: '#/definitions/os.File'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
      website:
        type: string
    type: object
  models.HotelGrade:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      hotel_type:
        $ref: '#/definitions/models.HotelType'
      hotel_type_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.HotelType:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      grades:
        items:
          $ref: '#/definitions/models.HotelGrade'
        type: array
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.Province:
    properties:
      alias:
        type: string
      cities:
        items:
          $ref: '#/definitions/models.City'
        type: array
      country:
        $ref: '#/definitions/models.Country'
      country_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.Room:
    properties:
      clean_status:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      max_beds:
        type: integer
      name:
        type: string
      room_type:
        $ref: '#/definitions/models.RoomType'
      room_type_id:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.RoomType:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      hotel:
        $ref: '#/definitions/models.Hotel'
      hotel_id:
        type: integer
      id:
        type: integer
      max_guest_count:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.User:
    properties:
      address:
        type: string
      birth_date:
        type: string
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
      username:
        type: string
    type: object
  os.File:
    type: object
info:
  contact: {}
  description: Swagger documentation for reservation API .
  license:
    name: MIT
  title: Reservation API
  version: "1.0"
paths:
...
  /room-types:
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RoomType'
            type: array
      summary: findAll RoomTypes
      tags:
      - RoomType
    post:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: RoomType
        in: body
        name: RoomType
        required: true
        schema:
          $ref: '#/definitions/models.RoomType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoomType'
      summary: crete RoomType
      tags:
      - RoomType
  /room-types/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Delete RoomType
      tags:
      - RoomType
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoomType'
      summary: find RoomType by id
      tags:
      - RoomType
    put:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      - description: RoomType
        in: body
        name: RoomType
        required: true
        schema:
          $ref: '#/definitions/models.RoomType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoomType'
      summary: update RoomType
      tags:
      - RoomType
  /rooms:
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Room'
            type: array
      summary: findAll Rooms
      tags:
      - Room
  /rooms/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Delete Room
      tags:
      - Room
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Room'
      summary: find Room by id
      tags:
      - Room
    put:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      - description: Room
        in: body
        name: Room
        required: true
        schema:
          $ref: '#/definitions/models.Room'
      - description: Room
        in: body
        name: Room
        required: true
        schema:
          $ref: '#/definitions/models.Room'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Room'
      summary: update Room
      tags:
      - Room
...
swagger: "2.0"